Fix Teams availability client bundle boundary
